Create fin_fd.py
This file uses an explicit finite difference scheme with central differences to obtain the temperature distribution and heat flow of a single extended surface, or fin. The exact value of the heat flow is also obtained for comparison. 

I used an example provided by The Kitchin Research Group at Carnegie Mellon University as a reference for the execution of the finite difference scheme.