Create single_stack.py
This file uses a hybrid analytical/numerical technique to obtain the temperature distribution and heat flows in a single stack finned array. This is essentially a single base fin with two edge fins attached to its tip. The junction temperature is obtained analytically by using the technique described in Kraus and Kern (1972) and is subsequently used as a boundary condition to finish the calculations numerically.