Create double_stack.py
This program obtains the temperature profile and heat flows of a double stack finned array in a similar method to the single stack array. This time, there are two junction temperatures that are obtained instead of one. Once those are calculated, the rest of the problem is calculated numerically. Fin tips are all adiabatic as in the previous cases.